Address code-review findings on the Stage 1 sharing-node foundation
Correctness:
- shares: scanner now follows symlinked files and folders (metadata() instead
  of file_type(), which reflected the raw dir entry and silently skipped
  symlinks), with a canonical-path cycle guard so a symlink to an ancestor
  can't loop. Adds a symlink test.
- peer_message: FolderContentsResponse is modeled as a directory LIST and
  decodes all directories instead of dropping every one past the first; encode
  still matches Nicotine+'s single-directory output for the common case.
- peer_message: UserInfoResponse picture is read as one bounded bulk copy
  (Reader::bytes) with a MAX_PICTURE_LEN cap, instead of an unbounded
  byte-by-byte loop.

Efficiency:
- search_response: partial-word matching does one pass over the word index
  (bucketing per partial) instead of a full scan per partial term.
- search_response: create_file_info_list partial-sorts (select_nth) to keep the
  max_results smallest-by-path candidates instead of fully sorting all matches,
  and skips phrase lowercasing entirely when there are no excluded phrases.
- shares: per-file token de-dup via the bucket's last id (no per-file HashSet or
  token clone).
- uploads: dropped the redundant queued_order vector; FIFO order is derived from
  ids (a transfer id is its insertion order), so dequeue is no longer an O(n)
  scan+shift and there's one less structure to keep consistent.

Cleanup / robustness:
- uploads: enqueue no longer takes an unused virtual_path.
- shares: add_virtual indexes one entry without the filesystem; the matcher
  tests use it instead of hand-rolling indexing that omitted the real dedup.
- config: max_search_results is u32 (was usize) for cross-platform config.

160 cargo tests pass; clippy clean (only pre-existing warnings); bazel green.

+++ crates/soulrust/src/config.rs
@@ -88,8 +88,10 @@ pub struct SharingConfig {
     pub fifo_queue: bool,
     /// Whether to answer incoming searches with our shared files.
     pub respond_to_searches: bool,
-    /// Hard cap on files returned for a single incoming search.
-    pub max_search_results: usize,
+    /// Hard cap on files returned for a single incoming search. `u32` (not
+    /// `usize`) so a config written on one platform parses identically on
+    /// another.
+    pub max_search_results: u32,
 }
 
 impl Default for SharingConfig {
